DataRepository

+++ src/main/java/org/example/ws/service/GreetingService.java
@@ -1,6 +1,5 @@
 package org.example.ws.service;
 
-import java.math.BigInteger;
 import java.util.Collection;
 
 import org.example.ws.model.Greeting;
@@ -9,7 +8,7 @@ public interface GreetingService {
 	
 	Collection<Greeting> findAll();
 	
-	Greeting findOne(BigInteger id);
+	Greeting findOne(Long id);
 	
 	Greeting create(Greeting greeting);
